requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.2
bcrypt==5.0.0
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
dill==0.4.0
distlib==0.4.0
dnspython==2.8.0
docker==7.1.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
testcontainers==4.14.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==16.0
wrapt==2.0.1
